+++ RealUI_Bugs/Libs/!BugGrabber/locales.lua
@@ -14,14 +14,14 @@ function addon:LoadTranslations(locale, L)
 	if locale == "koKR" then
 L["ADDON_CALL_PROTECTED"] = "[%s] 애드온 '%s' 보호된 함수 호출 '%s'."
 L["ADDON_CALL_PROTECTED_MATCH"] = "^%[(.*)%] (애드온 '.*' 보호된 함수 호출 '.*'.)$"
-L["ADDON_DISABLED"] = "|cffffff00!BugGrabber|r와 %s|1은;는; 함께 사용할 수 없습니다; %s|1은;는; 강제로 중지되었습니다. 만약 당신이 원하면, 접속을 종료한 후, !BugGrabber를 비활성하고 %s|1을;를; 활성화하세요.|r"
-L["BUGGRABBER_STOPPED"] = "|cffffff00당신의 UI에 매우 많은 오류가 있습니다. 결과적으로, 당신의 게임 환경은 타락했습니다. 이 메시지를 다시 보지 않으려면 오류가 있는 애드온을 비활성하거나 업데이트하세요.|r"
+L["ADDON_DISABLED"] = "|cffffff00!BugGrabber|r와 %s|1은;는; 함께 사용할 수 없습니다. 따라서 %s|1은;는; 강제로 중지되었습니다. 사용하고 싶다면 접속 종료 후, !BugGrabber를 끄고 %s|1을;를; 사용하세요.|r"
+L["BUGGRABBER_STOPPED"] = "|cffffff00UI에 오류가 너무 많습니다. 결과적으로 게임 환경이 타락했습니다. 이 메시지를 다시 보지 않으려면 오류가 있는 애드온을 사용 중지하거나 업데이트하세요.|r"
 L["ERROR_DETECTED"] = "%s |cffffff00수집됨, 자세한 정보는 링크를 클릭하세요.|r"
-L["ERROR_UNABLE"] = "|cffffff00!BugGrabber는 다른 플레이어의 오류는 발견할 수 없습니다. 이런 기능을 지원하는 BugSack이나 비슷한 디스플레이 애드온을 설치해주세요.|r"
-L["NO_DISPLAY_1"] = "|cffffff00당신은 !BugGrabber를 디스플레이 애드온 없이 실행한 것 같습니다. !BugGrabber는 게임 오류 확인을 위한 슬래시 명령어를 제공하고 있지만, 표시 애드온은 당신이 더 편리한 방법으로 이러한 오류를 관리할 수 있게 도와줍니다.|r"
-L["NO_DISPLAY_2"] = "|cffffff00표준 디스플레이는 BugSack으로 불리며, 아마도 당신은 !BugGrabber를 발견한 동일 사이트에서 찾을 수 있습니다.|r"
-L["NO_DISPLAY_STOP"] = "|cffffff00만약 당신이 이것에 대해 다시 떠올리고 싶지 않다면, /stopnag 를 실행하세요.|r"
-L["STOP_NAG"] = "|cffffff00다음 패치때까지 !BugGrabber는 디스플레이 애드온이 없는 것에 대해 성가시게 하지 않습니다.|r"
+L["ERROR_UNABLE"] = "|cffffff00!BugGrabber는 혼자서 다른 플레이어의 오류를 검색할 수 없습니다. 이 기능을 제공할 수 있는 BugSack이나 비슷한 디스플레이 애드온을 설치해주세요.|r"
+L["NO_DISPLAY_1"] = "|cffffff00당신은 !BugGrabber를 표시 애드온 없이 실행한 것 같습니다. !BugGrabber는 게임 오류 확인을 위한 슬래시 명령어를 제공하고 있지만, 표시 애드온은 당신이 더 편리한 방법으로 이러한 오류를 관리할 수 있게 도와줍니다.|r"
+L["NO_DISPLAY_2"] = "|cffffff00표준 디스플레이는 BugSack이라고 하며, 아마도 !BugGrabber를 구한 동일 사이트에서 찾을 수 있습니다.|r"
+L["NO_DISPLAY_STOP"] = "|cffffff00만약 이에 대해 다시 알림받고 싶지 않다면, /stopnag를 실행하세요.|r"
+L["STOP_NAG"] = "|cffffff00다음 패치때까지 !BugGrabber는 표시 애드온이 없는 것에 대해 성가시게 하지 않습니다.|r"
 L["USAGE"] = "|cffffff00사용법: /buggrabber <1-%d>.|r"
 
 	elseif locale == "deDE" then
@@ -92,14 +92,14 @@ L["USAGE"] = "|cffffff00Использование: /buggrabber <1-%d>.|r"
 	elseif locale == "frFR" then
 L["ADDON_CALL_PROTECTED"] = "[%s] L'AddOn '%s' a tenté d'appeler la fonction protégée '%s'."
 L["ADDON_CALL_PROTECTED_MATCH"] = "^%[(.*)%] (L'AddOn '.*' a tenté d'appeler la fonction protégée '.*'.)$"
-L["ADDON_DISABLED"] = "|cffffff7fBugGrabber|r et |cffffff7f%s|r ne peuvent pas être lancés en même temps. |cffffff7f%s|r a été désactivé. Si vous le souhaitez, vous pouvez vous déconnecter, désactiver |cffffff7fBugGrabber|r et réactiver |cffffff7f%s|r." -- Needs review
-L["BUGGRABBER_STOPPED"] = "|cffffff7fBugGrabber|r a cessé de capturer des erreurs, car plus de %d erreurs ont été capturées par seconde. La capture sera reprise dans %d secondes." -- Needs review
-L["NO_DISPLAY_1"] = "|cffff4411Vous ne semblez pas utiliser !BugGrabber avec un add-on d'affichage. Bien que les erreurs enregistrées par !BugGrabber soient accessibles par ligne de commande, un add-on d'affichage peut vous aidez à gérer ces erreurs plus aisément.|r" -- Needs review
-L["NO_DISPLAY_2"] = "|cffff4411L'add-on d'affichage originel s'appelle |r|cff44ff44BugSack|r|cffff4411, vous devriez pouvoir le trouver sur le même site que !BugGrabber.|r" -- Needs review
+L["ADDON_DISABLED"] = "|cffffff7fBugGrabber|r et |cffffff7f%s|r ne peuvent pas être lancés en même temps. |cffffff7f%s|r a été désactivé. Si vous le souhaitez, vous pouvez vous déconnecter, désactiver |cffffff7fBugGrabber|r et réactiver |cffffff7f%s|r."
+L["BUGGRABBER_STOPPED"] = "|cffffff7fBugGrabber|r a cessé de capturer des erreurs, car plus de %d erreurs ont été capturées par seconde. La capture sera reprise dans %d secondes."
+L["NO_DISPLAY_1"] = "|cffff4411Vous ne semblez pas utiliser !BugGrabber avec un add-on d'affichage. Bien que les erreurs enregistrées par !BugGrabber soient accessibles par ligne de commande, un add-on d'affichage peut vous aidez à gérer ces erreurs plus aisément.|r"
+L["NO_DISPLAY_2"] = "|cffff4411L'add-on d'affichage originel s'appelle |r|cff44ff44BugSack|r|cffff4411, vous devriez pouvoir le trouver sur le même site que !BugGrabber.|r"
 L["NO_DISPLAY_STOP"] = [=[|cffff4411Pour ne plus voir ce rappel, utiliser la commande |cff44ff44/stopnag|r|cffff4411.|r
-]=] -- Needs review
-L["STOP_NAG"] = "|cffff4411!BugGrabber ne vous rappellera plus l'existence de |r|cff44ff44BugSack|r|cffff4411 jusqu'à la prochaine mise à jour.|r" -- Needs review
-L["USAGE"] = "Utilisation: /buggrabber <1-%d>." -- Needs review
+]=]
+L["STOP_NAG"] = "|cffff4411!BugGrabber ne vous rappellera plus l'existence de |r|cff44ff44BugSack|r|cffff4411 jusqu'à la prochaine mise à jour.|r"
+L["USAGE"] = "Utilisation: /buggrabber <1-%d>."
 
 	elseif locale == "esMX" then
 L["ADDON_CALL_PROTECTED"] = "[%s] El addon '%s' ha intentado llamar a la función protegida '%s'."
@@ -115,17 +115,17 @@ L["STOP_NAG"] = "|cffff4411BugGrabber no te recordará sobre el desaparecido |r|
 L["USAGE"] = "Uso: /buggrabber <1-%d>"
 
 	elseif locale == "ptBR" then
-L["ADDON_CALL_PROTECTED"] = "[%s] O AddOn '%s' tentou chamar a função protegida '%s'."
+L["ADDON_CALL_PROTECTED"] = "[%s] O Addon '%s' tentou chamar a função protegida '%s'."
 L["ADDON_CALL_PROTECTED_MATCH"] = "^%[(.*)%] (AddOn '.*' tentou chamar a função protegida '.*'.)$"
-L["ADDON_DISABLED"] = "|cffffff7fBugGrabber|r e |cffffff7f%s|r não podem existir juntos. |cffffff7f%s|r foi desabilitado por causa disso. Se você quiser, você pode sair, desabilitar o |cffffff7fBugGrabber|r e habilitar o |cffffff7f%s|r." -- Needs review
-L["BUGGRABBER_STOPPED"] = "|cffffff7fBugGrabber|r parou de capturar erros, já que capturou mais de %d erros por segundo. A captura será resumida em %d segundos." -- Needs review
-L["ERROR_DETECTED"] = "%s |cffffff00capturado, clique no link para mais informações.|r" -- Needs review
-L["ERROR_UNABLE"] = "|cffffff00!BugGrabber não pode receber erros por outros jogadores por si só. Por favor, instale BugSack ou outro programa que oferece essa funcionalidade." -- Needs review
-L["NO_DISPLAY_1"] = "|cffff4411Aparentemente você está usando !BugGrabber sem nenhum addon de exibição para acompanhá-lo. Apesar do !BugGrabber fornecer um comando para acessar os erros dentro do jogo, um addon de exibição pode ajudar você a gerenciar esses erros de uma forma mais conveniente.|r" -- Needs review
-L["NO_DISPLAY_2"] = "|cffff4411O exibidor padrão do !BugGrabber é conhecido por |r|cff44ff44BugSack|r|cffff4411, e pode provavelmente ser encontrado no mesmo site onde você achou o !BugGrabber.|r" -- Needs review
-L["NO_DISPLAY_STOP"] = "|cffff4411Se você não deseja ser lembrado disso novamente, por favor utilize o comando |cff44ff44/stopnag|r|cffff4411.|r" -- Needs review
-L["STOP_NAG"] = "|cffff4411!BugGrabber não irá perturbar sobre não ter detectado o |r|cff44ff44BugSack|r|cffff4411 até a próxima atualização.|r" -- Needs review
-L["USAGE"] = "Uso: /buggraber <1-%d>" -- Needs review
+L["ADDON_DISABLED"] = "|cffffff00!BugGrabber e %s não podem coexistirem; %s tem que ser desativado. Se você quiser, desconecte, desative o !BugGrabber, e ative %s.|r"
+L["BUGGRABBER_STOPPED"] = "|cffffff00Existem muitos erros na sua interface. Como resultado, a experiência com o jogo pode ser desagradável. Desative ou atualize os Addons com falhas, se você não quiser ver essa mensagem novamente.|r"
+L["ERROR_DETECTED"] = "%s |cffffff00capturado, clique no link para mais informações.|r "
+L["ERROR_UNABLE"] = "|cffffff00!BugGrabber, por si só, é incapaz de receber erro de outros jogadores. Por favor. instale o BugSack ou outro programa que oferece esta funcionalidade.|r"
+L["NO_DISPLAY_1"] = "|cffffff00Parece que você está usando o !BugGrabber sem nenhum addon para visualização. Embora o haja um comando interno para acessar os relatórios de erros, um complemento possa ser necessário para gerenciar esses erros de forma mais conveniente.|r"
+L["NO_DISPLAY_2"] = "|cffffff00A ferramenta de exibição padrão é chamada BugSack, e provavelmente, você encontrará no mesmo site que você encontrou o !BugGrabber.|r"
+L["NO_DISPLAY_STOP"] = "|cffffff00Se você não quiser ser lembrado disto novamente, utilize o comando /stopnag.|r"
+L["STOP_NAG"] = "|cffffff00!BugGrabber não comentará sobre a ausência de uma ferramenta de exibição até a próxima versão.|r"
+L["USAGE"] = "|cffffff00Uso: /buggrabber <1-%d>.|r"
 
 	elseif locale == "itIT" then
 L["ADDON_CALL_PROTECTED"] = "[%s] AddOn '%s' ha cercato di chiamare la funzione protetta '%s'."