+++ RealUI_Chat/RealUI_Chat.lua
@@ -1,0 +1,6 @@
+--local ADDON_NAME, private = ...
+
+-- Lua Globals --
+-- luacheck: globals
+
+-- Libs --