+++ RealUI_Tooltips/RealUI_Tooltips.lua
@@ -1,0 +1,4 @@
+--local ADDON_NAME, private = ...
+
+-- Lua Globals --
+-- luacheck: globals